@@ -1,18 +1,25 @@
 [tool.poetry]
 name = "bpy-addon-build"
-version = "0.4.1"
+version = "0.5"
 description = "A build system to make building and testing Blender addons 10 times easier"
-authors = ["StandingPad"]
+authors = ["Mahid Sheikh <mahid@standingpad.org>"]
 license = "BSD-3-Clause"
 homepage = "https://github.com/StandingPadAnimations/bpy-build"
 
+packages = [
+    { include = "bpy_addon_build" },
+    { include = "lib_bpybuild_ext" },
+]
+
 [tool.poetry.dependencies]
-python = ">=3.8"
+python = ">=3.9"
 attrs = "^23.1.0"
 pyyaml = "^6.0.1"
 rich = "^13.7.0"
 typeguard = "^4.1.5"
 typing-extensions = "^4.11.0"
+tomli = "^2.0.1"
+packaging = "^24.1"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.4.1"
