@@ -10,9 +10,9 @@ from bpy_addon_build.args import Args
 from bpy_addon_build.config import Config
 
 INSTALL_PATHS: list[str] = [
-    "~/AppData/Roaming/Blender Foundation/Blender/{0}/scripts/addons",
-    "~/Library/Application Support/Blender/{0}/scripts/addons",
-    "~/.config/blender/{0}/scripts/addons",
+    "~/AppData/Roaming/Blender Foundation/Blender/",
+    "~/Library/Application Support/Blender/",
+    "~/.config/blender/",
 ]
 
 # Must be ignored because Mypy likes
