@@ -1,28 +1,39 @@
 from __future__ import annotations
 
-import sys
 import traceback
+from dataclasses import field
 from decimal import Decimal, getcontext
+from pathlib import Path
 from typing import Dict, List, Literal, Optional, TypedDict
 
 from attrs import frozen
 from rich.console import Console
 from typing_extensions import NotRequired, Union
 
-from .util import EXIT_FAIL, check_string, print_error
+from .util import check_string, exit_fail, print_error
 
+# Base settings
 ADDON_FOLDER: Literal["addon_folder"] = "addon_folder"
 BUILD_NAME: Literal["build_name"] = "build_name"
+BUILD_EXTENSION: Literal["build_extension"] = "build_extension"
 INSTALL_VERSIONS: Literal["install_versions"] = "install_versions"
+
+# Actions
 BUILD_ACTIONS: Literal["build_actions"] = "build_actions"
 SCRIPT: Literal["script"] = "script"
 IGNORE_FILTERS: Literal["ignore_filters"] = "ignore_filters"
+DEPENDS_ON: Literal["depends_on"] = "depends_on"
+
+# Extension Settings
+EXTENSION_SETTINGS: Literal["extension_settings"] = "extension_settings"
+BUILD_LEGACY: Literal["build_legacy"] = "build_legacy"
+REMOVE_BL_INFO: Literal["remove_bl_info"] = "remove_bl_info"
 
 VERSION_JUMPS = {
     "2.83": Decimal(2.9),
     "2.93": Decimal(3.0),
     "3.6": Decimal(4.0),
-    "3.60": Decimal(4.0), # Include version with 0
+    "3.60": Decimal(4.0),  # Include version with 0
 }
 
 
@@ -31,6 +42,15 @@ class BuildActionDict(TypedDict):
 
     script: NotRequired[str]
     ignore_filters: NotRequired[list[str]]
+    depends_on: NotRequired[list[str]]
+
+
+class ExtensionSettingsDict(TypedDict):
+    """TypeDict verson of ExtensionSettings"""
+
+    build_legacy: NotRequired[bool]
+    build_name: NotRequired[str]
+    remove_bl_info: NotRequired[bool]
 
 
 class ConfigDict(TypedDict):
@@ -38,6 +58,8 @@ class ConfigDict(TypedDict):
 
     addon_folder: str
     build_name: str
+    build_extension: bool
+    extension_settings: NotRequired[ExtensionSettingsDict]
     install_versions: NotRequired[list[Union[float, str]]]
     build_actions: NotRequired[dict[str, Optional[BuildActionDict]]]
 
@@ -61,6 +83,40 @@ class BuildAction:
 
     script: Optional[str] = None
     ignore_filters: Optional[List[str]] = None
+    depends_on: Optional[list[str]] = None
+
+
+BUILT_IN_ACTIONS_FOLDER = Path(__file__).parent.joinpath("built_in_actions")
+BUILT_IN_ACTS = {
+    "extension": BuildAction(str(BUILT_IN_ACTIONS_FOLDER.joinpath("extension.py")))
+}
+
+
+# Must be ignored to pass Mypy as this has
+# an expression of Any, likely due to how
+# attrs works
+@frozen  # type: ignore
+class ExtensionSettings:
+    """Class storing all settings for Blender extensions
+
+    Attributes
+    ----------
+    build_legacy: bool
+        Whether to build a legacy addon or not
+
+    build_name: Optional[str]
+        Build name for the built extension. Useful when
+        building both a legacy addon and Blender extension
+
+    remove_bl_info: bool
+        Whether to remove bl_info or not from an addon
+
+        Note: this is set True if build_legacy is False
+    """
+
+    build_legacy: bool
+    build_name: Optional[str]
+    remove_bl_info: bool
 
 
 # Must be ignored to pass Mypy as this has
@@ -78,6 +134,12 @@ class Config:
     build_name: str
         Name of the final build
 
+    build_extension: bool
+        Whether to build a Blender 4.2+ extension
+
+    extension_settings: Optional[ExtensionSettings]
+        Settings for building an extension
+
     versions: Optional[List[float]]
         List of Blender versions to install the final addon to
 
@@ -87,8 +149,11 @@ class Config:
 
     addon_folder: str
     build_name: str
+    build_extension: bool = True
+    extension_settings: Optional[ExtensionSettings] = None
     install_versions: Optional[List[Decimal]] = None
     build_actions: Optional[Dict[str, BuildAction]] = None
+    additional_actions: list[str] = field(default_factory=list)
 
 
 def build_config(data: ConfigDict) -> Config:
@@ -108,6 +173,8 @@ def build_config(data: ConfigDict) -> Config:
 
     console = Console()
     parsed_build_acts: dict[str, BuildAction] = {}
+    additional_actions: list[str] = []
+    parsed_extension_settings: Optional[ExtensionSettings] = None
     install_versions: list[Decimal] = []
 
     # Set the precision for Decimal to
@@ -120,7 +187,7 @@ def build_config(data: ConfigDict) -> Config:
     try:
         if ADDON_FOLDER not in data:
             print_error("addon_folder not defined!", console)
-            sys.exit(EXIT_FAIL)
+            exit_fail()
 
         # Disallow '.' as a folder option
         # as it's been found to cause issues
@@ -133,17 +200,52 @@ def build_config(data: ConfigDict) -> Config:
         # As such, this is simply not allowed.
         elif data[ADDON_FOLDER] == ".":
             print_error("Addon must be in a subfolder!", console)
-            sys.exit(EXIT_FAIL)
+            exit_fail()
         elif not check_string(data[ADDON_FOLDER]):
             print_error("addon_folder uses unsupported characters!", console)
-            sys.exit(EXIT_FAIL)
+            exit_fail()
 
         if BUILD_NAME not in data:
             print_error("build_name must be defined!", console)
-            sys.exit(EXIT_FAIL)
+            exit_fail()
         elif not check_string(data[BUILD_NAME]):
             print_error("build_name uses unsupported characters!", console)
-            sys.exit(EXIT_FAIL)
+            exit_fail()
+
+        if BUILD_EXTENSION in data and data[BUILD_EXTENSION]:
+            parsed_build_acts["extension"] = BUILT_IN_ACTS["extension"]
+            additional_actions.append("extension")
+            if EXTENSION_SETTINGS in data:
+                extension_settings_data = data[EXTENSION_SETTINGS]
+                if (
+                    REMOVE_BL_INFO in extension_settings_data
+                    and BUILD_LEGACY not in extension_settings_data
+                ):
+                    print_error(
+                        "Cannot set extension_settings::remove_bl_info if legacy builds are not performed!",
+                        console,
+                    )
+                    exit_fail()
+                if BUILD_NAME in extension_settings_data and not check_string(
+                    extension_settings_data[BUILD_NAME]
+                ):
+                    print_error(
+                        "extension_settings::build_name uses unsupported characters!",
+                        console,
+                    )
+                    exit_fail()
+                parsed_extension_settings = ExtensionSettings(
+                    build_legacy=extension_settings_data[BUILD_LEGACY]
+                    if BUILD_LEGACY in extension_settings_data
+                    else False,
+                    build_name=extension_settings_data[BUILD_NAME]
+                    if BUILD_NAME in extension_settings_data
+                    else None,
+                    remove_bl_info=extension_settings_data[REMOVE_BL_INFO]
+                    if REMOVE_BL_INFO in extension_settings_data
+                    and extension_settings_data[BUILD_LEGACY]
+                    else False,
+                )
 
         if INSTALL_VERSIONS in data:
             for ver in data[INSTALL_VERSIONS]:
@@ -156,13 +258,22 @@ def build_config(data: ConfigDict) -> Config:
                     install_versions += version_shorthand_expand(ver)
                 else:
                     print_error(f"{ver} isn't a valid floating point value", console)
-                    sys.exit(EXIT_FAIL)
+                    exit_fail()
+
+            if BUILD_EXTENSION in data and data[BUILD_EXTENSION]:
+                greater_4_2 = [x for x in install_versions if x >= Decimal("4.2")]
+                if not len(greater_4_2):
+                    print_error(
+                        "When building extensions, 4.2 must be included in install_versions!",
+                        console,
+                    )
+                    exit_fail()
 
         if BUILD_ACTIONS in data:
             for act in data[BUILD_ACTIONS]:
                 if not check_string(act):
                     print_error(f"{act} uses unsupported characters!", console)
-                    sys.exit(EXIT_FAIL)
+                    exit_fail()
                 action_data = data[BUILD_ACTIONS][act]
                 if action_data is not None:
                     # We need to make sure the script name
@@ -177,7 +288,7 @@ def build_config(data: ConfigDict) -> Config:
                             f"Script defined for {act} uses unsupported characters in file name!",
                             console,
                         )
-                        sys.exit(EXIT_FAIL)
+                        exit_fail()
 
                     # Add the action to parsed_build_acts to
                     # use later in Config construction
@@ -186,27 +297,33 @@ def build_config(data: ConfigDict) -> Config:
                         ignore_filters=action_data[IGNORE_FILTERS]
                         if IGNORE_FILTERS in action_data
                         else None,
+                        depends_on=action_data[DEPENDS_ON]
+                        if DEPENDS_ON in action_data
+                        else None,
                     )
                     continue
 
                 # If an action has nothing defined, what's the
                 # point of said action?
                 print_error(f"{act} must have something defined!", console)
-                sys.exit(EXIT_FAIL)
+                exit_fail()
 
     except Exception as e:
         console.print(e)
         console.print(traceback.format_exc())
         console.print(data)
-        sys.exit(EXIT_FAIL)
+        exit_fail()
 
     return Config(
         addon_folder=data["addon_folder"],
         build_name=data["build_name"],
+        build_extension=data[BUILD_EXTENSION] if BUILD_EXTENSION in data else False,
+        extension_settings=parsed_extension_settings,
         install_versions=sorted(install_versions, reverse=True)
         if "install_versions" in data
         else None,
         build_actions=parsed_build_acts if len(parsed_build_acts) else None,
+        additional_actions=additional_actions,
     )
 
 
